+++ djstripe/models/payment_methods.py
@@ -142,14 +142,16 @@ class LegacySourceMixin:
 
         self.delete()
 
-    def api_retrieve(self, api_key=None):
+    def api_retrieve(self, api_key=None, stripe_account=None):
         # OVERRIDING the parent version of this function
         # Cards & Banks Accounts must be manipulated through a customer or account.
         # TODO: When managed accounts are supported, this method needs to check if
         # either a customer or account is supplied to determine the
         # correct object to use.
         api_key = api_key or self.default_api_key
-        customer = self.customer.api_retrieve(api_key=api_key)
+        customer = self.customer.api_retrieve(
+            api_key=api_key, stripe_account=stripe_account
+        )
 
         # If the customer is deleted, the sources attribute will be absent.
         # eg. {"id": "cus_XXXXXXXX", "deleted": True}
@@ -158,6 +160,8 @@ class LegacySourceMixin:
             # like an invalid ID error.
             raise InvalidRequestError("No such source: %s" % (self.id), "id")
 
+        # This will retrieve the source using the account ID where the customer resides,
+        # so we don't have to pass `stripe_account`.
         return customer.sources.retrieve(self.id, expand=self.expand_fields)
 
 
@@ -378,7 +382,7 @@ class Source(StripeModel):
         null=True,
         blank=True,
         help_text=(
-            "Amount associated with the source. "
+            "Amount (as decimal) associated with the source. "
             "This is the amount for which the source will be chargeable once ready. "
             "Required for `single_use` sources."
         ),
@@ -572,3 +576,43 @@ class PaymentMethod(StripeModel):
             payment_method, customer=customer, **extra_kwargs
         )
         return cls.sync_from_stripe_data(stripe_payment_method)
+
+    def detach(self):
+        """
+        Detach the payment method from its customer.
+
+        :return: Returns true if the payment method was newly detached, \
+                 false if it was already detached
+        :rtype: bool
+        """
+        # Find customers that use this
+        customers = Customer.objects.filter(default_payment_method=self).all()
+        changed = True
+
+        # special handling is needed for legacy "card"-type PaymentMethods,
+        # since detaching them deletes them within Stripe.
+        # see https://github.com/dj-stripe/dj-stripe/pull/967
+        is_legacy_card = self.id.startswith("card_")
+
+        try:
+            self.sync_from_stripe_data(self.api_retrieve().detach())
+
+            # resync customer to update .default_payment_method and
+            # .invoice_settings.default_payment_method
+            for customer in customers:
+                Customer.sync_from_stripe_data(customer.api_retrieve())
+
+        except (InvalidRequestError,):
+            # The source was already detached. Resyncing.
+
+            if self.pk and not is_legacy_card:
+                self.sync_from_stripe_data(self.api_retrieve())
+            changed = False
+
+        if self.pk:
+            if is_legacy_card:
+                self.delete()
+            else:
+                self.refresh_from_db()
+
+        return changed